+++ pkg/rule/retry/retry_errors.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package retryablerule
+package retry
 
 import (
 	"regexp"
