+++ pkg/rule/retry/retryable_rule.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package retryablerule
+package retry
 
 import (
 	"context"
